Move generateUser() and helper f(x) from scripts.js to domUpdates obj
